requirement2.txt
aiofiles==22.1.0
aiohttp==3.7.4.post0
aiosignal==1.3.1
aiostream==0.4.5
altair==4.2.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.3.4
asttokens==2.1.0
async-timeout==3.0.1
attrs==21.2.0
audioread==3.0.0
autopep8==2.0.0
av==10.0.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.1
clip
commonmark==0.9.1
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
docarray==0.18.1
docker==6.0.1
docopt==0.6.2
elastic-apm==6.2.1
elasticsearch==7.13.1
entrypoints==0.4
executing==1.2.0
fastapi==0.65.1
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
googleapis-common-protos==1.56.4
grpcio==1.47.2
grpcio-health-checking==1.47.2
grpcio-reflection==1.47.2
h11==0.12.0
httptools==0.5.0
huggingface-hub==0.0.12
idna==3.2
importlib-metadata==5.0.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jcloud==0.0.36
jedi==0.18.1
jina==3.4.2
jina-hubble-sdk==0.22.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.2
jupyter_client==7.4.6
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kaggle==1.5.12
kiwisolver==1.4.4
librosa==0.9.1
llvmlite==0.39.1
lz4==3.1.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==5.1.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.4
opencv-python==4.6.0.66
opentelemetry-api==1.14.0
opentelemetry-exporter-otlp==1.14.0
opentelemetry-exporter-otlp-proto-grpc==1.14.0
opentelemetry-exporter-otlp-proto-http==1.14.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation==0.35b0
opentelemetry-instrumentation-aiohttp-client==0.35b0
opentelemetry-instrumentation-asgi==0.35b0
opentelemetry-instrumentation-fastapi==0.35b0
opentelemetry-instrumentation-grpc==0.35b0
opentelemetry-proto==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-semantic-conventions==0.35b0
opentelemetry-util-http==0.35b0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.4
pooch==1.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==3.20.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.8.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
python-slugify==7.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.4.24
requests==2.28.1
resampy==0.4.2
rich==12.6.0
sacremoses==0.0.53
scikit-learn==1.1.3
scipy==1.8.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.11.0
soupsieve==2.3.2.post1
stack-data==0.6.1
starlette==0.14.2
streamlit==1.12.0
terminado==0.17.0
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.10.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.0+cu117
torchvision==0.14.0+cu117
tornado==6.2
tqdm==4.64.0
traitlets==5.5.0
transformers==4.9.1
typing-extensions==3.10.0.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.5
uvicorn==0.14.0
uvloop==0.17.0
validators==0.20.0
watchdog==2.1.9
watchfiles==0.18.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
webvtt-py==0.4.6
widgetsnbextension==4.0.3
wrapt==1.14.1
yarl==1.6.3
zipp==3.10.0
